--- data/sopletter/src/main/java/org/sopt/official/data/sopletter/dto/response/SopletterDefaultMessagesResponseDto.kt
@@ -1,58 +1,0 @@
-/*
- * MIT License
- * Copyright 2026 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official.data.sopletter.dto.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SopletterDefaultMessagesResponseDto(
-    @SerialName("topicId")
-    val topicId: Long,
-    @SerialName("title")
-    val title: String,
-    @SerialName("totalCount")
-    val totalCount: Int,
-    @SerialName("nextCursor")
-    val nextCursor: Long? = null,
-    @SerialName("hasNext")
-    val hasNext: Boolean,
-    @SerialName("messages")
-    val messages: List<SopletterMessageDto>,
-)
-
-@Serializable
-data class SopletterMessageDto(
-    @SerialName("messageId")
-    val messageId: Long,
-    @SerialName("previewContent")
-    val previewContent: String,
-    @SerialName("colorCode")
-    val colorCode: String,
-    @SerialName("rotationDegree")
-    val rotationDegree: Double,
-    @SerialName("shapeType")
-    val shapeType: String,
-)
