+++ data/sopletter/src/main/java/org/sopt/official/data/sopletter/mapper/SopletterMapper.kt
@@ -23,12 +23,12 @@
  * SOFTWARE.
  */
 package org.sopt.official.data.sopletter.mapper
-import org.sopt.official.data.sopletter.dto.response.SopletterDefaultMessagesResponseDto
+import org.sopt.official.data.sopletter.dto.response.SopletterMessagesResponseDto
 import org.sopt.official.data.sopletter.dto.response.SopletterMessageDetailResponseDto
 import org.sopt.official.data.sopletter.dto.response.SopletterMessageDto
 import org.sopt.official.data.sopletter.dto.response.SopletterTopicDto
 import org.sopt.official.data.sopletter.dto.response.SopletterTopicsResponseDto
-import org.sopt.official.domain.sopletter.model.SopletterDefaultMessages
+import org.sopt.official.domain.sopletter.model.SopletterMessages
 import org.sopt.official.domain.sopletter.model.SopletterMessageDetail
 import org.sopt.official.domain.sopletter.model.SopletterMessage
 import org.sopt.official.domain.sopletter.model.SopletterShapeType
@@ -36,7 +36,7 @@ import org.sopt.official.domain.sopletter.model.SopletterTopic
 
 internal fun SopletterTopicsResponseDto.toDomain(): List<SopletterTopic> = topics.map(SopletterTopicDto::toDomain)
 
-internal fun SopletterDefaultMessagesResponseDto.toDomain(): SopletterDefaultMessages = SopletterDefaultMessages(
+internal fun SopletterMessagesResponseDto.toDomain(): SopletterMessages = SopletterMessages(
     topicId = topicId,
     title = title,
     totalCount = totalCount,
