+++ data/sopletter/src/main/java/org/sopt/official/data/sopletter/repository/SopletterRepositoryImpl.kt
@@ -28,7 +28,7 @@ import javax.inject.Inject
 import org.sopt.official.common.coroutines.suspendRunCatching
 import org.sopt.official.data.sopletter.datasource.SopletterDataSource
 import org.sopt.official.data.sopletter.mapper.toDomain
-import org.sopt.official.domain.sopletter.model.SopletterDefaultMessages
+import org.sopt.official.domain.sopletter.model.SopletterMessages
 import org.sopt.official.domain.sopletter.model.SopletterMessageDetail
 import org.sopt.official.domain.sopletter.model.SopletterTopic
 import org.sopt.official.domain.sopletter.repository.SopletterRepository
@@ -39,7 +39,7 @@ internal class SopletterRepositoryImpl @Inject constructor(
     override suspend fun getDefaultMessages(
         cursor: Long?,
         size: Int,
-    ): Result<SopletterDefaultMessages> = suspendRunCatching {
+    ): Result<SopletterMessages> = suspendRunCatching {
         sopletterDataSource.getDefaultMessages(
             cursor = cursor,
             size = size,
