+++ data/sopletter/src/main/java/org/sopt/official/data/sopletter/service/SopletterService.kt
@@ -25,7 +25,7 @@
 package org.sopt.official.data.sopletter.service
 
 import org.sopt.official.data.sopletter.dto.request.UpdateSopletterMessageRequestDto
-import org.sopt.official.data.sopletter.dto.response.SopletterDefaultMessagesResponseDto
+import org.sopt.official.data.sopletter.dto.response.SopletterMessagesResponseDto
 import org.sopt.official.data.sopletter.dto.response.SopletterMessageDetailResponseDto
 import org.sopt.official.data.sopletter.dto.response.SopletterReportFormResponseDto
 import org.sopt.official.data.sopletter.dto.response.SopletterTopicsResponseDto
@@ -42,7 +42,7 @@ interface SopletterService {
     suspend fun getDefaultMessages(
         @Query("cursor") cursor: Long?,
         @Query("size") size: Int,
-    ): SopletterDefaultMessagesResponseDto
+    ): SopletterMessagesResponseDto
 
     @GET("sopt-letter/report-form")
     suspend fun getReportForm(): SopletterReportFormResponseDto
