--- domain/sopletter/src/main/java/org/sopt/official/domain/sopletter/model/SopletterDefaultMessages.kt
@@ -1,54 +1,0 @@
-/*
- * MIT License
- * Copyright 2026 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official.domain.sopletter.model
-
-data class SopletterDefaultMessages(
-    val topicId: Long,
-    val title: String,
-    val totalCount: Int,
-    val nextCursor: Long?,
-    val hasNext: Boolean,
-    val messages: List<SopletterMessage>,
-)
-
-data class SopletterMessage(
-    val messageId: Long,
-    val previewContent: String,
-    val colorCode: String,
-    val rotationDegree: Double,
-    val shapeType: SopletterShapeType,
-)
-
-enum class SopletterShapeType {
-    SMOOTH,
-    SHARP,
-    POINT,
-    CLOUD;
-
-    companion object {
-        fun from(value: String): SopletterShapeType =
-            entries.find { it.name == value } ?: SHARP
-    }
-}
