+++ domain/sopletter/src/main/java/org/sopt/official/domain/sopletter/model/SopletterMessages.kt
@@ -1,0 +1,30 @@
+package org.sopt.official.domain.sopletter.model
+
+data class SopletterMessages(
+    val topicId: Long,
+    val title: String,
+    val totalCount: Int,
+    val nextCursor: Long?,
+    val hasNext: Boolean,
+    val messages: List<SopletterMessage>,
+)
+
+data class SopletterMessage(
+    val messageId: Long,
+    val previewContent: String,
+    val colorCode: String,
+    val rotationDegree: Double,
+    val shapeType: SopletterShapeType,
+)
+
+enum class SopletterShapeType {
+    SMOOTH,
+    SHARP,
+    POINT,
+    CLOUD;
+
+    companion object {
+        fun from(value: String): SopletterShapeType =
+            entries.find { it.name == value } ?: SHARP
+    }
+}
