+++ domain/sopletter/src/main/java/org/sopt/official/domain/sopletter/repository/SopletterRepository.kt
@@ -24,7 +24,7 @@
  */
 package org.sopt.official.domain.sopletter.repository
 
-import org.sopt.official.domain.sopletter.model.SopletterDefaultMessages
+import org.sopt.official.domain.sopletter.model.SopletterMessages
 import org.sopt.official.domain.sopletter.model.SopletterMessageDetail
 import org.sopt.official.domain.sopletter.model.SopletterTopic
 
@@ -32,7 +32,7 @@ interface SopletterRepository {
     suspend fun getDefaultMessages(
         cursor: Long? = null,
         size: Int = DEFAULT_PAGE_SIZE,
-    ): Result<SopletterDefaultMessages>
+    ): Result<SopletterMessages>
 
     suspend fun getReportFormUrl(): Result<String>
